move generate_package_guid to wix.rs
